% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeExtents}
\alias{computeExtents}
\title{Find the extents in the given data.frame}
\usage{
computeExtents(df, ncpus = 1, minsupport = 0)
}
\arguments{
\item{df}{A data.frame with the forma concept data}

\item{ncpus}{The number of cpu cores to be used for parallel processing}

\item{minsupport}{The minumum support threshold}
}
\value{
A list containing the extents found in the given context
}
\description{
This function call the pcbo algorithm to find the concepts in the context 
represented in the dataframe passed as a parameter.
}
